Learn how to use avrdude tool
The avrdude tool will be helpful for later when the RPi goes headless.
I figure the sooner I learn how to use it, the better.

This commit contains a simple arduino code to test the avrdude tool
with as well as a script that acts as a wrapper around the avrdude
command. This command was copied directly from the arduino IDE,
and modified so that other file locations may be inserted into the
upload command.

Author: Trenton Foster
Issue: New Feature