Update upload script
Updated the script that uploads a hex file from the RPi to the arduino
so that the user is requried to put the loction of the .hex file to be
uploaded on the same line as the script call.
Later, this will allow for multiple arduino projects to be used in one
large project.

Author: Trenton Foster
Issue: Code Improvement