Servo addition and bug fix
Through testing with my group it was found that the solenoid is not
a great device solution to lock and unlock the mailbox. So, we've switched
to a servo implementation instead.

Also, I found an error in auth.sh involving the path in open_mailbox().
This bug was fixed in this commit.

Type: Bug Fix and New Feature
Author: Trenton Foster